Simplify and clean up ninja-file
Previously a custom variable for the build directory was used instead
of the standard one (`$build_dir` vs `$builddir`). This caused the
`.ninja_log` to be generated in the repository root instead of the
`build/`-directory. Therefore this commit renames that variable.

Furthermore, this commit removes the `create-build-directory` rule
and invocation of that rule, since ninja creates the necessary direc-
tories by itself.